Print count of a substring in a given string
It prints the number of occurance of a given substring in a string.
Program takes the source string and the string to be searched as
argumnets.

it contain 2 function as subString which return the substring from the
source string and another function called subStringFrequency which check
the returned substring is equal to the substring to be searched.